instrument: restructure some tests, add bad clipboard data test
